feat(helix-ui): add Helix LLM configuration forms to DemoDataPage and DemoSetupPanel
- Add state variables for Helix config (base_url, api_key, environment_id, agent_id)
- Implement fetchHelixStatus() and handleHelixSave() handlers for both components
- Add Helix configuration section to DemoDataPage with 4 input fields, status pill, and action buttons
- Mirror Helix section to DemoSetupPanel for /configure?tab=demo-management access
- Append CSS classes (.demo-data-helix-status, .demo-data-helix-fields, .demo-data-helix-hint)
- Fix form label accessibility with htmlFor attributes

+++ banking_api_server/routes/mcpExchangeMode.js
@@ -9,7 +9,7 @@ const router = express.Router();
 
 router.get('/exchange-mode', (req, res) => {
   if (!req.session?.user) return res.status(401).json({ error: 'not_authenticated' });
-  const mode = req.session.mcpExchangeMode || 'single';
+  const mode = req.session.mcpExchangeMode || 'double';
   res.json({ mode });
 });
